Adding xml docs for Cake.RepoVersion
Also adds a devcontainer.json file
