+++ src/repo-version/repo-version.csproj
@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp2.1</TargetFramework>
     <RootNamespace>repo_version</RootNamespace>
     <PackAsTool>true</PackAsTool>
     <ToolCommandName>repo-version</ToolCommandName>
